requirements.txt
cffi==1.13.2
Click==7.0
docutils==0.15.2
Jinja2==2.9.6
jmespath==0.9.4
MarkupSafe==1.1.1
pycparser==2.19
pygit2==0.28.2
python-dateutil==2.8.0
six==1.13.0
urllib3==1.25.7
awscli==1.16.309
botocore==1.13.45
arrow==0.15.4
feedgen==0.8.0
dataclasses_json==0.3.6
boto3==1.10.45
